update: split Results component
